messaging_app/messaging_app/urls.py
from django.contrib import admin
from django.http import JsonResponse
from django.urls import path, include
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
    TokenVerifyView,
)

urlpatterns = [
    path("admin/", admin.site.urls),

    # JWT Auth
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/token/verify/', TokenVerifyView.as_view(), name='token_verify'),

    # DRF browsable login/logout
    path('api-auth/', include('rest_framework.urls')),
    path("", lambda request: JsonResponse({"message": "Welcome to the Messaging API!"})),

    # Your app’s API
    path('api/', include('chats.urls')),
]
