messaging_app/requirements.txt
-------- -------
# Django Messaging App project dependencies

----------------------------- -------
asgiref                       3.9.1
autopep8                      2.3.2
Django                        5.2.4
django-cors-headers           4.7.0
django-environ                0.12.0
django-filter                 25.1
djangorestframework           3.16.0
djangorestframework_simplejwt 5.5.1
drf-nested-routers            0.94.2
Faker                         37.4.2
mysqlclient                   2.2.7
pip                           25.1.1
pycodestyle                   2.14.0
PyJWT                         2.10.1
sqlparse                      0.5.3
tzdata                        2025.2